Remove redundent function in agent namespace(2/n)
